pyproject.toml
[project]
name = "rps-conversational-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.6",
    "langgraph>=1.0.7",
]
